https://github.com/scoverage/scalac-scoverage-plugin/pull/109 introduced code redundance in report writer classes. Base class created and common code moved there. "IOUtils.relativeSource" moved there too, because it's reports specific and works on paths as strings, not real files.
