+++ scalac-scoverage-plugin/src/main/scala/scoverage/report/ScoverageXmlWriter.scala
@@ -2,20 +2,17 @@ package scoverage.report
 
 import java.io.File
 
-import _root_.scoverage._
+import scoverage._
 
 import scala.xml.{Node, PrettyPrinter}
 
 /** @author Stephen Samuel */
-class ScoverageXmlWriter(sourceDirectories: Seq[File], outputDir: File, debug: Boolean) {
+class ScoverageXmlWriter(sourceDirectories: Seq[File], outputDir: File, debug: Boolean) extends BaseReportWriter(sourceDirectories, outputDir) {
 
   def this (sourceDir: File, outputDir: File, debug: Boolean) {
     this(Seq(sourceDir), outputDir, debug);
   }
 
-  // Source paths in canonical form WITH trailing file separator
-  val formattedSourcePaths: Seq[String] = sourceDirectories filter ( _.isDirectory ) map ( _.getCanonicalPath + File.separator )
-  
   def write(coverage: Coverage): Unit = {
     val file = IOUtils.reportFile(outputDir, debug)
     IOUtils.writeToFile(file, new PrettyPrinter(120, 4).format(xml(coverage)))
@@ -103,7 +100,4 @@ class ScoverageXmlWriter(sourceDirectories: Seq[File], outputDir: File, debug: B
     </package>
   }
 
-  private def relativeSource(src: String): String = IOUtils.relativeSource(src, formattedSourcePaths)
-
 }
-
